more work on the cost analysis, mostly finished. need costs for type signature parsing
